+++ src/storybook-helpers/knobs/forwardAction.ts
@@ -2,9 +2,9 @@ import { action, ActionOptions } from '@storybook/addon-actions';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-export function forwardAction<T extends any[]>(name: string, callback: (...args: T) => void, options?: ActionOptions) {
+export function forwardAction<T extends any[]>(name: string, callback: (...args: T) => any, options?: ActionOptions) {
   return (...args: T) => {
     action(name, options)(...args);
-    callback(...args);
+    return callback(...args);
   };
 }